+++ endpoints/summary.go
@@ -35,12 +35,12 @@ func HandleSummary(logger *zap.SugaredLogger, summaryStore storage.SummaryStore)
 			)
 
 			// Report the error to the client.
-			code := int64(500)
+			code := 500
 			resp := &api.TerseSummaryDefault{Payload: &models.Error{
-				Code:    code,
+				Code:    int64(code),
 				Message: message,
 			}}
-			resp.SetStatusCode(int(code))
+			resp.SetStatusCode(code)
 			return resp
 		}
